refactor: 함수를 분리한다.
