+++ app/src/main/java/woowacourse/shopping/di/DependencyInjector.kt
@@ -4,8 +4,8 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import woowacourse.shopping.SingletonObject
 import kotlin.reflect.KClass
+import kotlin.reflect.KParameter
 import kotlin.reflect.full.companionObject
 import kotlin.reflect.full.companionObjectInstance
 import kotlin.reflect.full.isSubclassOf
@@ -15,22 +15,20 @@ inline fun <reified T : Any> inject(): T {
     val primaryConstructor = T::class.primaryConstructor
         ?: throw Throwable("주 생성자를 찾을 수 없습니다.")
 
-    val arguments = primaryConstructor
-        .parameters
-        .map { it.type.classifier as KClass<*> }
-        .map {
-            if (it.companionObject?.isSubclassOf(SingletonObject::class) == true) {
-                (it.companionObjectInstance as SingletonObject<*>).get()
-            } else {
-                it.primaryConstructor?.call()
-            }
-        }
-        .toTypedArray()
-
-    return primaryConstructor.call(*arguments)
+    return primaryConstructor.call(*primaryConstructor.parameters.instantiate())
 }
 
-inline fun <reified T : ViewModel> getViewModelFactory(): ViewModelProvider.Factory =
+fun List<KParameter>.instantiate(): Array<Any?> =
+    map { (it.type.classifier as KClass<*>).instantiate() }.toTypedArray()
+
+fun KClass<*>.instantiate(): Any? =
+    if (companionObject?.isSubclassOf(SingletonObject::class) == true) {
+        (companionObjectInstance as SingletonObject<*>).get()
+    } else {
+        primaryConstructor?.call()
+    }
+
+inline fun <reified T : ViewModel> getInjectedViewModelFactory(): ViewModelProvider.Factory =
     viewModelFactory {
         initializer {
             inject<T>()
