+++ app/src/main/java/woowacourse/shopping/ui/MainActivity.kt
@@ -8,14 +8,14 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.ActivityMainBinding
-import woowacourse.shopping.di.getViewModelFactory
+import woowacourse.shopping.di.getInjectedViewModelFactory
 import woowacourse.shopping.ui.cart.CartActivity
 
 class MainActivity : AppCompatActivity() {
 
     private val binding by lazy { ActivityMainBinding.inflate(layoutInflater) }
 
-    private val viewModel: MainViewModel by viewModels { getViewModelFactory<MainViewModel>() }
+    private val viewModel: MainViewModel by viewModels { getInjectedViewModelFactory<MainViewModel>() }
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
